Create run_kubernetes function (#37)
We need to be able to run the model on kubernetes. This PR creates a python function `run_kubernetes` for doing so. It has not yet been added to the bash `fv3run` command.

Fixes #33.

Other API changes:

- `DependencyError` is removed, instead a `DelayedImportError` object is used to delay the `ImportError` until the code attempts to access a member of the imported module.

Non-API/internal changes:

- pytest is now used to run tests
- `make coverage` now uses pytest
- python 3 versions of python binaries are called explicitly (e.g. `python3` and `pip3`)
- internally the package now uses gcsfs instead of gsutil for copy operations. `fsspec`, which has the same API as `gscfs`, is now used for local copy operations other than linking
- added an example for creating a run directory using an external path, which I used to manually test remote copy operations
- asset list generation for remote and local directories has been unified into one function
- tree walking responsibility and asset creation responsibility of asset list generation has been separated into two functions
- distinction is made between "get" and "copy" in file operation helper function names
- native fv3run functionality uses a named logger instead of the root logger
- removed unused copy/link methods in `_datastore.py`
- fixed some code-formatting in sphinx docs

Commit history:

* initial commit of run endpoint files

* use non-root user to execute in docker

* refactor code to use one global variable for cache location

* implemented routines to set cache location

* moved data folder to testdata

* update testdata dirname in test

* fixed forcingdata tests to use temporary data directory

* fix path resolution for options to use dynamic cache dir

* Add testdata to manifest

* set stacksize to unlimited, bind mount output directory

* document fv3run

* clear up logic for _google_cloud_copy a bit

* rename run_experiment to run, run module to fv3run

* Add gcp-gcr to circleci environment

* add fv3run entry point, use dict instead of Namelist internally

* added pytest to dev requirements

* use internal refresh routine in test

* use pytest to run tests in circleci

* clean up run directory after tests, enable all fv3run tests

* enable all fv3run tests

* enable docker

* Added unit tests for docker arg construction

* added gcsfs to requirements_dev.txt

* use environment image for circleci

* switch bach to python base image

* disable mpi on FileNotFoundError for mpirun

* use cwd and not test directory for determining absolute paths from relative

* pull fv3gfs-python image for tests

* properly label run command

* use gcloud auth to configure docker auth

* add google cloud sdk service component image

* use orb gcr-auth command

* use machine executor so we can pull docker image

* remove prefix from gcr-auth

* remove setup-remote-docker because machine

* revert to docker image based method

* disable docker test on circleci, can still run it locally

* remove unused gcr orb

* test circleci docker version

* disable docker if image is unavailable

* reset cache dir when finished

* reset cache dir to original at end of tests

* remove subprocess from non-mpi tests as it always uses mpi

* specify defaults are for namelist

* clean up old comments

* remove print statements

* restructured fv3run into multiple files

* intention revealing comment

* add documentation for python run function

* fix documentation formatting

* create cache dir if it does not exist

* add optional requirement for fv3run

* initial kubernetes code

* remove unused code

* fixed tests, responded to review comments

* updated usage instructions for fv3run with another example

* added missing parentheses to docs

* renamed _get_cache_subdir to _get_internal_cache_dir

* add initial kubernetes run function to public namespace

* remove comment

* make output directory before docker run command

* add --rm flag to docker call

* split public run function for native and docker

* fixed bugs while getting example to work, removed job management

* replaced gsutil usage with gcsfs

* centralize gcloud utilities

* add google_cloud_project option

* fixes of gcloud functionality

* create restart directory in run directory

* specify image pull policy, set needed MPI flags

* Add prefix to dirname when comparing atainst prefixed location using relpath

* fix variable names in config dict copy operation

* refactor asset generation into two functions for readability

* add mpi flags to reference command

* use logger instead of logging module

* add parentheses to function call

* refactor run_kubernetes to use a _submit_job helper

* remove unused file operations

* rename copy_directory to put_directory to reflect usage

* rename gcloud.py to filesystem.py since it handles local ops also

* added docstring to run_kubernetes

* fix gcloud name to filesystem

* update documentation with kubernetes job submission details

* fix tests

* add tests for kubernetes job creation, fix code to pass

* add kubernetes to dev requirements

* fix image name

* Update docs/usage.rst


* fix bugs in fv3config.run_kubernetes example

* add kubernetes as an extras_require entry

* use links to API wherever functions appear

* remove print statements

* added missing comma between requirements

* remove google_cloud_project environment variable

* remove _is_gsbucket_path for _is_local_path, reduce usage of this info

* remove explicit google-cloud-storage dependency

* refactor secret tests into their own helper

* rename _list variables to be consistent with kubernetes api

* fix coverage to use pytest

+++ fv3config/_asset_list.py
@@ -1,15 +1,13 @@
-import logging
 import os
-import shutil
-from subprocess import check_call, Popen, PIPE
 from ._datastore import (
     get_initial_conditions_directory, get_orographic_forcing_directory,
-    get_base_forcing_directory, is_gsbucket_url
+    get_base_forcing_directory
 )
 from ._tables import (
     get_data_table_filename, get_diag_table_filename, get_field_table_filename
 )
-from ._exceptions import ConfigError, DependencyError
+from ._exceptions import ConfigError
+from . import filesystem
 
 
 def is_dict_or_list(option):
@@ -102,72 +100,51 @@ def get_asset_dict(source_location, source_name, target_location='',
     return asset
 
 
-def asset_list_from_path(path, target_location='', copy_method='copy'):
-    """Return an asset_list corresponding to all files within path"""
-    if is_gsbucket_url(path):
-        if gsutil_is_installed():
-            return asset_list_from_gs_bucket(path, target_location=target_location)
-        else:
-            raise DependencyError(f'Optional dependency gsutil not found. Files in {path} will not be copied to {target_location}.')
+def _without_dot(path):
+    if path == '.':
+        return ''
     else:
-        return asset_list_from_local_dir(path,
-                                         target_location=target_location,
-                                         copy_method=copy_method)
+        return path
 
 
-def asset_list_from_local_dir(source_directory, target_location='', copy_method='copy'):
-    """Return asset_list from all files in the local path source_directory. Will
-    recurse to subdirectories within source_directory.
+def asset_list_from_path(from_location, target_location='', copy_method='copy'):
+    """Return asset_list from all files within a given path.
 
     Args:
-        source_directory (str): path to local directory from which to generate
-            asset_list.
+        location (str): local path or google cloud storage url.
         target_location (str, optional): target_location used for generated assets.
             Defaults to '' which is root of run-directory.
-        copy_method (str, optional): copy_method used for generated assets. Defaults
-            to 'copy'.
-
-    Returns:
-        list: a list of asset dictionaries
-    """
-    asset_list = []
-    for root, dirs, files in os.walk(source_directory):
-        if root == source_directory:
-            file_target_location = target_location
-        else:
-            file_target_location = os.path.join(target_location,
-                                                os.path.basename(root))
-        for file in files:
-            asset_list.append(get_asset_dict(os.path.join(source_directory, root),
-                                             file,
-                                             target_location=file_target_location,
-                                             copy_method=copy_method))
-    return asset_list
-
-
-def asset_list_from_gs_bucket(url, target_location=''):
-    """Return asset_list from all files that begin with a google cloud storage url.
-    Will not recurse to files that are in "sub-directories" of the url.
-
-    Args:
-        url (str): google cloud storage url from which to generate asset_list.
-        target_location (str, optional): target_location used for generated assets.
-            Defaults to '' which is root of run-directory.
+        copy_method ('copy' or 'link', optional): whether to copy or link assets,
+            defaults to 'copy'. If location is a google cloud storage url, this option
+            is ignored and files are copied.
 
     Returns:
         list: a list of asset dictionaries
         """
+    if not filesystem._is_local_path(from_location):
+        copy_method = 'copy'
     asset_list = []
-    # TODO: use gcsfs instead of gsutil for following code
-    stdout_str = Popen(['gsutil', 'ls', url], stdout=PIPE).stdout.read()
-    path_list = stdout_str.decode().split('\n')[:-1]
-    for path in path_list:
-        dirname, basename = os.path.split(path)
-        if basename != '':
-            asset_list.append(get_asset_dict(dirname,
-                                             basename,
-                                             target_location=target_location))
+    for dirname, basename, relative_target_location in _asset_walk(from_location):
+        asset_list.append(
+            get_asset_dict(
+                dirname,
+                basename,
+                target_location=os.path.join(target_location, relative_target_location),
+                copy_method=copy_method
+            )
+        )
     return asset_list
+
+
+def _asset_walk(location):
+    fs = filesystem.get_fs(location)
+    protocol_prefix = filesystem._get_protocol_prefix(location)
+    path_list = fs.walk(location)
+    for dirname, _, files in path_list:
+        dirname = protocol_prefix + dirname
+        subdir_target_location = os.path.relpath(dirname, start=location)
+        for basename in files:
+            yield dirname, basename, _without_dot(subdir_target_location)
 
 
 def write_asset(asset, target_directory):
@@ -184,13 +161,18 @@ def write_asset(asset, target_directory):
     if not os.path.exists(os.path.dirname(target_path)):
         os.makedirs(os.path.dirname(target_path))
     if copy_method == 'copy':
-        copy_file(source_path, target_path)
+        get_file(source_path, target_path)
     elif copy_method == 'link':
         link_file(source_path, target_path)
     else:
         raise ConfigError(
             f'Behavior of copy_method {copy_method} not defined for {source_path} asset'
         )
+
+
+def write_assets_to_directory(config, target_directory):
+    asset_list = config_to_asset_list(config)
+    write_asset_list(asset_list, target_directory)
 
 
 def write_asset_list(asset_list, target_directory):
@@ -221,28 +203,23 @@ def config_to_asset_list(config):
         if is_dict_or_list(config['patch_files']):
             asset_list += ensure_is_list(config['patch_files'])
         else:
-            raise ConfigError('patch_files item in config dictionary must be a dict or list')
+            raise ConfigError(
+                'patch_files item in config dictionary must be an asset dict or '
+                'list of asset dicts')
     return asset_list
 
 
 def link_file(source_item, target_item):
+    if any(not filesystem._is_local_path(item) for item in [source_item, target_item]):
+        raise NotImplementedError(
+            'cannot perform linking operation involving remote urls '
+            f'from {source_item} to {target_item}'
+        )
     if os.path.exists(target_item):
         os.remove(target_item)
-    os.link(source_item, target_item)
+    os.symlink(source_item, target_item)
 
 
-def copy_file(source_path, target_path):
-    if is_gsbucket_url(source_path):
-        if gsutil_is_installed():
-            check_call(['gsutil', 'cp', source_path, target_path])
-        else:
-            raise DependencyError(f'Optional dependency gsutil not found. File {source_path} will not be copied to {target_path}')
-    else:
-        shutil.copy(source_path, target_path)
-
-
-def gsutil_is_installed():
-    if shutil.which('gsutil') is None:
-        return False
-    else:
-        return True
+def get_file(source_path, target_path):
+    fs = filesystem.get_fs(source_path)
+    fs.get(source_path, target_path)
